% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{birth}
\alias{birth}
\title{Birth}
\usage{
birth(x, i, ...)
}
\arguments{
\item{x}{the object which determines method dispatch}

\item{i}{the index of the species undergoing birth}
}
\description{
Generic and methods for computing the birth process
}
