% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleParams.R
\name{buildFun}
\alias{buildFun}
\title{buildFun}
\usage{
buildFun(p)
}
\arguments{
\item{p}{something magical}
}
\value{
p stretched to niter
}
\description{
helper that, given a single value, builds a function
}
