% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolePhylo.R
\name{checkRolePhylo}
\alias{checkRolePhylo}
\title{check RolePhylo
checker function for validation}
\usage{
checkRolePhylo(object)
}
\arguments{
\item{object}{a supposed rolePhylo}
}
\value{
t/f if it's a rolePhylo
}
\description{
check RolePhylo
checker function for validation
}
