% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{class}
\name{comm-class}
\alias{comm-class}
\alias{localComm-class}
\alias{metaComm-class}
\title{S4 classes to specify community objects}
\description{
The \code{comm} class is used as the parent class for
\code{localComm} and \code{metaComm} child classes.
}
\section{Slots}{

\describe{
\item{\code{abundance}}{a numeric vector of abundances for each species}

\item{\code{traits}}{matrix of traits; the first column specifies the species index
and the subsequent columns specify the trait values}

\item{\code{Smax}}{a single integer specifying the total number of species ever
recorded in the local community (both locally extinct and extant)}

\item{\code{pi}}{a numeric vector of genetic diversities for each species}
}}

