% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dataFunCpp}
\alias{dataFunCpp}
\title{dataFunCpp}
\arguments{
\item{fun_name}{fun_name}

\item{local}{local}

\item{meta}{meta}

\item{phylo}{phylo}

\item{params}{params}

\item{niter}{niter}

\item{i}{i}

\item{dead_index}{dead_index}

\item{parent_indv}{parent_indv}

\item{dispersed_this_year}{dispersed_this_year}

\item{speciation_sp}{speciation_sp}
}
\description{
dataFunCpp
}
