% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{death}
\alias{death}
\title{Death}
\usage{
death(x, i, ...)
}
\arguments{
\item{x}{the object which determines method dispatch}

\item{i}{the index of the species undergoing death}
}
\description{
Generic and methods for computing the death process
}
