% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{.birthLocal}
\alias{.birthLocal}
\title{function to implement birth for \code{localComm} class objects}
\usage{
.birthLocal(x, i)
}
\arguments{
\item{x}{an object of class \code{localComm}}

\item{i}{the vector of indices of the species undergoing birth}
}
\description{
function to implement birth for \code{localComm} class objects
}
