% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.R
\name{.birthRole}
\alias{.birthRole}
\title{function to implement birth for \code{roleModel} class objects}
\usage{
.birthRole(x)
}
\arguments{
\item{x}{an object of class \code{roleModel}}
}
\description{
function to implement birth for \code{roleModel} class objects
}
