% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{.deathComm}
\alias{.deathComm}
\title{function to implement death for \code{*comm} class objects}
\usage{
.deathComm(x, i)
}
\arguments{
\item{x}{an object of class \code{localComm}}

\item{i}{the index of the species undergoing death}
}
\description{
function to implement death for \code{*comm} class objects
}
