% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{.deathPhylo}
\alias{.deathPhylo}
\title{function to implement death for \code{*rolePhylo} class objects}
\usage{
.deathPhylo(x, i)
}
\arguments{
\item{x}{an object of class \code{rolePhylo}}

\item{i}{the index of the tip undergoing death}
}
\description{
function to implement death for \code{*rolePhylo} class objects
}
