% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{.deathRole}
\alias{.deathRole}
\title{function to implement death for \code{*roleModel} class objects}
\usage{
.deathRole(x)
}
\arguments{
\item{x}{an object of class \code{roleModel}}
}
\description{
function to implement death for \code{*roleModel} class objects
}
