% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-divStats.R
\name{.hillDivSAD}
\alias{.hillDivSAD}
\title{HillDivSAD
function for abundance-based}
\usage{
.hillDivSAD(n, q)
}
\arguments{
\item{n}{is a vector of species abundances}

\item{q}{hill order}
}
\description{
HillDivSAD
function for abundance-based
}
