% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-divStats.R
\name{.hillDivTrait}
\alias{.hillDivTrait}
\title{HillDivTrait
function for trait-based}
\usage{
.hillDivTrait(X, q)
}
\arguments{
\item{X}{a matrix with first column = abund, second column = traits}

\item{q}{order for hill number}
}
\description{
HillDivTrait
function for trait-based
}
