% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immigration.R
\name{.immigrationLocal}
\alias{.immigrationLocal}
\title{function to implement immigration for \code{localComm} class objects}
\usage{
.immigrationLocal(x, i)
}
\arguments{
\item{x}{an object of class \code{localComm}}

\item{i}{the vector of indices of the species undergoing immigration}
}
\description{
function to implement immigration for \code{localComm} class objects
}
