% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immigration.R
\name{.immigrationRole}
\alias{.immigrationRole}
\title{function to implement immigration for \code{roleModel} class objects}
\usage{
.immigrationRole(x)
}
\arguments{
\item{x}{an object of class \code{roleModel}}
}
\description{
function to implement immigration for \code{roleModel} class objects
}
