% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleModel.R
\name{.lseriesFromSN}
\alias{.lseriesFromSN}
\title{logseriesfromsn}
\usage{
.lseriesFromSN(S, N)
}
\arguments{
\item{S}{number of species}

\item{N}{number of individuals}
}
\description{
non-exported helper function to solve for parameter of logseries
}
