% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciation.R
\name{.specRole}
\alias{.specRole}
\title{function to implement speciation for \code{roleModel} class objects}
\usage{
.specRole(x)
}
\arguments{
\item{x}{an object of class \code{roleModel}}
}
\description{
function to implement speciation for \code{roleModel} class objects
}
