% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalState.R
\docType{methods}
\name{getFinalState}
\alias{getFinalState}
\alias{getFinalState,roleModel-method}
\alias{getFinalState,roleExperiment-method}
\title{Get the final state of a model run.}
\usage{
getFinalState(x, modID, ...)

\S4method{getFinalState}{roleModel}(x)

\S4method{getFinalState}{roleExperiment}(x, modID)
}
\arguments{
\item{x}{The `roleExperiment` or `roleModel` object to retrieve the final state from.}

\item{modID}{Optional argument to retrieve the final state for specific model(s) in a
`roleExperiment` object; can be a vector of unique IDs.}

\item{...}{passed}
}
\description{
Gets the final state of a model from either a `roleExperiment` 
    or a `roleModel`.
}
\details{
If `modID` is not supplied and `x` is a `roleExperiment`, then a 
    list of `roleData` objects for each model run is returned.
}
