% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumStats.R
\name{getSumStats}
\alias{getSumStats}
\alias{getSumStats,roleData-method}
\alias{getSumStats,roleModel-method}
\alias{getSumStats,roleExperiment-method}
\title{Get summary statistics for RoLE objects.}
\usage{
getSumStats(x, funs = .getDefaultSumStatsFuns(), moreArgs, ...)

\S4method{getSumStats}{roleData}(x, funs, moreArgs)

\S4method{getSumStats}{roleModel}(x, funs, moreArgs)

\S4method{getSumStats}{roleExperiment}(x, funs, moreArgs)
}
\arguments{
\item{x}{The object to calculate summary statistics on.}

\item{funs}{A named list of functions to calculate the summary statistics; can be a 
named list with a single function or many functions, but must be a 
named list of functions. If unspecified, defaults to including several standard stats. 
Included diversity index functions in roleR are: `hillAbund`, 
`hillGenetic`,`hillPhylo` & `richness`. Included raw stats functions are `rawAbundance`,
`rawSpAbundance`,`rawSppId`,`rawTraits`,`rawGenDiv`,`rawSeqs`,
`rawBranchLengths` & `rawApePhylo`.}

\item{moreArgs}{An optional named list of additional arguments to pass to the
functions listed in `funs`. If given, list names must match those in 
`funs`. Note: not all names in `funs` need to appear in `moreArgs`}

\item{...}{ignored}
}
\value{
data.frame containing summary stats where each row is a model snapshot
and each column is a summary stat requested by a function provided to `funs`
}
\description{
Applies different summary statistics functions to `roleExperiment`,
    `roleModel`, or `roleData` objects.
}
\details{
Users can define their own functions, so long as they work on any
    object of class `roleData`.
}
\examples{
# get the species richness 
# rich_stat <- getSumStats(model, funs = list(rich = richness))
# get many default summary stats
# stats <- getSumStats(model)

}
