% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRole.R
\name{getValuesFromParams}
\alias{getValuesFromParams}
\title{getValuesFromParams}
\usage{
getValuesFromParams(p, i)
}
\arguments{
\item{p}{an object of class `roleParams`}

\item{i}{the iterations over which to calculate parameter values}
}
\value{
a list of parameter vectors, each equal in length to `length(i)`
}
\description{
run iter functions over params to generate a new object of class paramValues 
that contains ONLY vectors of values and no functions
}
