% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillDiv.R
\name{hillDivSAD}
\alias{hillDivSAD}
\alias{hillDivTrait}
\title{Hill number diversities}
\usage{
hillDivSAD(x, q)

hillDivTrait(x, traits, q)
}
\arguments{
\item{x}{a vector of species abundances}

\item{q}{a vector of Renyei entropy scales}
}
\value{
A vector, same length as \code{q}, giving diversities for each
value of \code{q}
}
\description{
Hill number diversities for species abundance, trait, population
genetic, and phylogenetic data
}
\details{
Stub
}
\examples{
hillDivSAD(1:10, 0:3)
hillDivTrait(1:10, 1:10, 0:3)

}
