% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immigration.R
\name{immigration}
\alias{immigration}
\title{Immigration}
\usage{
immigration(x, i, ...)
}
\arguments{
\item{x}{the object which determines method dispatch}

\item{i}{the index of the species undergoing immigration}
}
\description{
Generic and methods for computing the immigration process
}
