% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iterModelCpp}
\alias{iterModelCpp}
\title{iterModelCpp}
\arguments{
\item{local}{local}

\item{meta}{meta}

\item{phylo}{phylo}

\item{params}{params}

\item{print}{print}
}
\description{
iterModelCpp
}
