% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iterSim}
\alias{iterSim}
\title{a function to iterate the RoLE model through \code{nstep} steps of simulation}
\arguments{
\item{an}{object of class \code{roleModelCpp} to iterate}

\item{nstep}{an \code{integer} of the number of steps to iterate}

\item{print}{a \code{bool} specifying whether to print the step outcomes}
}
\description{
a function to iterate the RoLE model through \code{nstep} steps of simulation
}
