% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iterSimCpp}
\alias{iterSimCpp}
\title{a C++ function to iterate the RoLE model through nstep steps of simulation}
\arguments{
\item{model}{the entire C++ model, an object of class roleModelCpp}

\item{nstep}{the number of steps to iterate}
}
\description{
a C++ function to iterate the RoLE model through nstep steps of simulation
}
