% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localComm.R
\docType{class}
\name{localComm-class}
\alias{localComm-class}
\title{An S4 class to specify the local community of the RoLE model}
\description{
An S4 class to specify the local community of the RoLE model
}
\section{Slots}{

\describe{
\item{\code{abundance}}{a numeric vector of abundances for each species}

\item{\code{traits}}{matrix of traits; the first column specifies the species index
(i.e. the index of that species in the \code{abundance} vector) and the
second column specifies the trait value}

\item{\code{pi}}{a numeric vector of genetic diversities for each species}

\item{\code{Smax}}{a single integer specifying the total number of species ever
recorded in the local community (both locally extinct and extant)}
}}

