% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{localCommCpp}
\alias{localCommCpp}
\title{a class to specify the local community}
\arguments{
\item{abundance_binary}{a numeric vector specifying the binary abundance (alive or dead) 
of each individual}

\item{traits}{a numeric vector of trait values for each individual}

\item{Imax}{a float specifying the total number of individuals ever
recorded in the local community (both locally extinct and extant)}

\item{pi}{a numeric vector of genetic diversities for each species}
}
\description{
a class to specify the local community
}
