% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{class}
\name{metaComm-class}
\alias{metaComm-class}
\alias{metaComm}
\title{The metacommunity of a `roleData` object}
\usage{
metaComm(spAbund, spTrait)
}
\arguments{
\item{spAbund}{spAbund}

\item{spTrait}{spTrait}
}
\description{
An S4 class to specify the state of a metacommunity
}
\section{Slots}{

\describe{
\item{\code{spAbund}}{a numeric vector of the relative abundances for each species in 
the meta community}

\item{\code{spTrait}}{a numeric vector of the average trait value for each species in 
the meta community}
}}

