% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metaCommCpp}
\alias{metaCommCpp}
\title{a class to specify the meta community}
\arguments{
\item{abundance}{a \code{numeric vector} of relative abundances for each species}

\item{traits}{a \code{numeric vector} of traits for each species}

\item{Smax}{a single \code{integer} specifying the total number of species ever
recorded in the local community (both locally extinct and extant)}
}
\description{
a class to specify the meta community
}
