% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-rawStats.R
\docType{methods}
\name{rawAbundance}
\alias{rawAbundance}
\alias{rawAbundance,roleData-method}
\alias{rawSppID}
\alias{rawSppID,roleData-method}
\alias{rawTraits}
\alias{rawTraits,roleData-method}
\alias{rawGenDiv}
\alias{rawGenDiv,roleData-method}
\alias{rawSeqs}
\alias{rawSeqs,roleData-method}
\alias{rawBranchLengths}
\alias{rawBranchLengths,roleData-method}
\alias{rawApePhylo}
\alias{rawApePhylo,roleData-method}
\title{Get raw data from `roleData` objects.}
\usage{
rawAbundance(x, ...)

\S4method{rawAbundance}{roleData}(x)

rawSppID(x, ...)

\S4method{rawSppID}{roleData}(x)

rawTraits(x, ...)

\S4method{rawTraits}{roleData}(x)

rawGenDiv(x, ...)

\S4method{rawGenDiv}{roleData}(x)

rawSeqs(x, ...)

\S4method{rawSeqs}{roleData}(x)

rawBranchLengths(x, ...)

\S4method{rawBranchLengths}{roleData}(x)

rawApePhylo(x, ...)

\S4method{rawApePhylo}{roleData}(x)
}
\arguments{
\item{x}{A `roleData` object.}

\item{...}{Additional parameters, ignored.}
}
\value{
A list with one element. Inside that one element is a vector of the
    raw data.
}
\description{
Gets raw data according to the function specification.
}
\details{
These functions return a list of length one (with summary statistics as a vector
    inside that list) because the summary statistics themselves can be variable
    length. This is distinct from other summary statistic functions which will
    always return a fixed length value.  When using raw summary statistic
    functions in `getSumStats` they will be returned in a list-column.
    rawAbundance gets the unordered abundance of species, 
    while rawSpAbundance gets the ordered abundance of species which is a much longer vector
}
