% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-rawStats.R
\docType{methods}
\name{rawSpAbundance}
\alias{rawSpAbundance}
\alias{rawSpAbundance,roleData-method}
\title{abundance}
\usage{
rawSpAbundance(x, ...)

\S4method{rawSpAbundance}{roleData}(x)
}
\arguments{
\item{x}{x}

\item{...}{ignored}
}
\description{
abundance

abundance on roleData
}
