% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleComm.R
\docType{class}
\name{roleComm-class}
\alias{roleComm-class}
\title{An S4 class to specify the entire hierrarchical community of the
RoLE model}
\description{
An S4 class to specify the entire hierrarchical community of the
RoLE model
}
\section{Slots}{

\describe{
\item{\code{localComm}}{an object of class \code{localComm}}

\item{\code{params}}{an object of calss \code{roleParams}}
}}

