% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData.R
\docType{class}
\name{roleData-class}
\alias{roleData-class}
\title{RoLE model data}
\description{
An S4 class to hold data from one timestep of one model run
}
\section{Slots}{

\describe{
\item{\code{localComm}}{an object of class \code{localComm}}

\item{\code{metaComm}}{an object of class \code{metaComm}}

\item{\code{phylo}}{an object of class \code{rolePhylo}}
}}

