% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleExperiment.R
\docType{class}
\name{roleExperiment-class}
\alias{roleExperiment-class}
\title{roleExperiment - one or more models bundled collectively}
\description{
An S4 class to represent a self-enclosed modeling experiment
}
\section{Slots}{

\describe{
\item{\code{modelRuns}}{a list of `roleData`objects}

\item{\code{allParams}}{a list of `roleParams` to use for each model}

\item{\code{info}}{a data.frame with model parameter information for each saved 
snapshot for all the models of the experiment}

\item{\code{inits}}{an optional list of initialized `roleModel` objects}
}}

