% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleExperiment-methods.R, R/roleExperiment.R
\docType{methods}
\name{[,roleExperiment,ANY,ANY,ANY-method}
\alias{[,roleExperiment,ANY,ANY,ANY-method}
\alias{[,roleExperiment-method}
\alias{$}
\alias{$,roleExperiment-method}
\alias{roleExperiment-class}
\alias{roleExperiment}
\title{Extract parts of a roleExperiment}
\usage{
\S4method{[}{roleExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{$}{roleExperiment}(x, name)

roleExperiment(allParams)
}
\arguments{
\item{x}{a roleExperiment}

\item{i}{i}

\item{j}{j}

\item{...}{ignored}

\item{drop}{default F}

\item{name}{thing to extract}

\item{allParams}{the list of model params to use for each model; can also be
`rolePriors` object}
}
\value{
a ready-to-run `roleExperiment`
}
\description{
An S4 class to represent a self-enclosed modeling experiment
}
\section{Slots}{

\describe{
\item{\code{modelRuns}}{a list of `roleData`objects}

\item{\code{allParams}}{a list of `roleParams` to use for each model}

\item{\code{info}}{a data.frame with model parameter information for each saved 
snapshot for all the models of the experiment}

\item{\code{inits}}{an optional list of initialized `roleModel` objects}
}}

\examples{
# create and run a roleExperiment with three different levels of dispersal 

}
