% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleModel.R
\docType{class}
\name{roleModel-class}
\alias{roleModel-class}
\title{A single RoLE model.}
\description{
An S4 class that holds one RoLE model realization. A model is 
    first initialized using a set of parameters, then run using those 
    parameters.
}
\details{
See the `roleR_intro` vignette for an example modeling workflow.
}
\section{Slots}{

\describe{
\item{\code{modelSteps}}{A list of `roleData` objects, one for each saved snapshot.
For example, the 3rd saved snapshot is accessed at `modelSteps[[3]]`.
Models that are not yet run only have one timestep, i.e. `modelSteps[[1]]`}

\item{\code{params}}{A `roleParams` object containing the params of the model}

\item{\code{info}}{a `data.frame` with one row for each saved snapshot (unrun models
will only have 1 row); columns are parameters, cells are parameter values 
at each snapshot}
}}

\examples{
# Create a model using a default set of params, then run it.
m <- roleModel(roleParams())
m <- runRole(m)

}
