% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleModel.R
\name{roleModel}
\alias{roleModel}
\title{Create a roleModel.}
\usage{
roleModel(params)
}
\arguments{
\item{params}{The params, an object of class `roleParams`, to use when the 
model is run.}
}
\value{
A ready-to-run `roleModel`.
}
\description{
Create a roleModel.
}
