% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleParams.R
\docType{class}
\name{roleParams-class}
\alias{roleParams-class}
\title{An S4 class to specify paramters of the RoLE model}
\description{
An S4 class to specify paramters of the RoLE model
}
\section{Slots}{

\describe{
\item{\code{params}}{a named list of numerical values of the parameters; if priors
are specified, \code{params} are used as initial values}

\item{\code{runType}}{a character vector of length one specifying the type of run
(either "sim" or "fit")}

\item{\code{priors}}{a names list of functions that generate samples from the
desired prior distributions}
}}

