% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleParams.R
\name{roleParams}
\alias{roleParams}
\title{Specify parameters for a RoLE model}
\usage{
roleParams(params, runType, priors)
}
\arguments{
\item{params}{a named list of parameter values, used as initial values if
priors are specified}

\item{runType}{can be either "sim" or "fit", specifying whether to run
simulations only, or also fit the model to data}

\item{priors}{a named list of functions to be used to generate samples from
the prior distributions of parameters}
}
\value{
an object of class \code{roleParams}
}
\description{
Specify parameters for a RoLE model
}
\seealso{
\code{\link{roleParams-class}}
}
