% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolePhylo.R
\docType{class}
\name{rolePhylo-class}
\alias{rolePhylo-class}
\alias{rolePhylo}
\title{A phylogeny of all the species of a `roleData`}
\usage{
rolePhylo(n, e, l, alive, tipNames, scale)
}
\arguments{
\item{n}{}

\item{e}{}

\item{l}{}

\item{alive}{}

\item{tipNames}{}
}
\value{
a `rolePhylo` object
}
\description{
An S4 class to specify a phylogeny for the purpose of the RoLE model
}
\section{Slots}{

\describe{
\item{\code{n}}{the number of tips in the phylogeny}

\item{\code{e}}{the numeric edge matrix of the phylogeny
Each row contains an ancestor-child pair where the 1st column is the ancestor and the 2nd is the child}

\item{\code{l}}{a numeric vector of edge lengths 
The units of l are the time steps (iterations) of the model
Each time step unit is equal to 1/J generations where J is the number of individuals in the local community}

\item{\code{alive}}{a logical vector indicating whether each tips is extant or not}

\item{\code{tipNames}}{a character vector of the names of each tip}

\item{\code{scale}}{a single numeric value of time scale translation to years}
}}

