\name{roleR-package}
\alias{roleR-package}
\alias{roleR}
\docType{package}
\title{
\packageTitle{roleR}
}
\description{
\packageDescription{roleR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{roleR}
\packageIndices{roleR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{roleR}

Maintainer: \packageMaintainer{roleR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
