% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleSim.R
\name{roleSim}
\alias{roleSim}
\title{RoLE model simulation}
\usage{
roleSim(params, nstep = NULL, nsim = 1)
}
\arguments{
\item{params}{list of parameters}

\item{nstep}{number of simulation steps to run; if \code{prop_equilib} is specified in the \code{params}
list, \code{nstep} will over-ride that parameter}

\item{nsim}{number of simulations to run}
}
\value{
An object of class \code{roleComm}
}
\description{
Simulate communities under the RoLE model
}
\details{
Stub
}
\examples{
params <- list(species_meta = 100,
               individuals_meta = 10000,
               individuals_local = 1000,
               dispersal_prob = 0.1,
               speciation_local = 0.01)
testSim <- roleSim(params, nstep = 10, nsim = 1)
testSim

}
