% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRole.R
\docType{methods}
\name{runRole}
\alias{runRole}
\alias{runRole,roleModel-method}
\alias{runRole,roleExperiment-method}
\title{Run a `roleModel` or `roleExperiment`}
\usage{
runRole(x, cores = 1)

\S4method{runRole}{roleModel}(x)

\S4method{runRole}{roleExperiment}(x, cores = 1)
}
\arguments{
\item{x}{`roleModel` or `roleExperiment` object to run}

\item{cores}{number of cores}
}
\value{
`roleModel` or `roleExperiment` run for the specified number of time
    steps
}
\description{
`runRole` takes a `role*` object and simulates all processes 
    specified in the given parameters
}
\details{
Prior to invoking `runRole`, the `role*` input object must be 
    initialized with a set of parameters, e.g., as generated by `roleParams`.
}
\examples{
# create and run a model
model <- roleModel(roleParams())
model <- runRole(model)

# create and run an experiment
p1 <- roleParams(dispersal_prob = 0.2)
p2 <- roleParams(dispersal_prob = 0.3)
p3 <- roleParams(dispersal_prob = 0.4)
exp <- roleExperiment(list(p1,p2,p3))
exp <- runRole(exp)
}
