% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciation.R
\name{speciation}
\alias{speciation}
\title{Speciation}
\usage{
speciation(x, ...)
}
\arguments{
\item{x}{the object which determines method dispatch}

\item{...}{additional parameters passed to specific methods}
}
\description{
Generic and methods for computing the speciation process
}
