% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectFunCpp}
\alias{vectFunCpp}
\title{vectFunCpp}
\usage{
vectFunCpp(fun_name, local, meta, phylo, params, niter, i)
}
\arguments{
\item{fun_name}{fun_name}

\item{local}{local}

\item{meta}{meta}

\item{phylo}{phylo}

\item{params}{params}

\item{niter}{niter}

\item{i}{i}
}
\description{
vectFunCpp
}
