% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRoLE.R
\name{writeRole}
\alias{writeRole}
\alias{writeRole,roleExperiment-method}
\title{Write a roleExperiment to disk}
\usage{
writeRole(x, directory = NULL, filename = NULL, save_txt = TRUE)

\S4method{writeRole}{roleExperiment}(x, directory = NULL, filename = NULL, save_txt = TRUE)
}
\arguments{
\item{x}{roleExperiment object to write.}

\item{directory}{directory (path) to write the object to.}

\item{filename}{name of the file to save.}

\item{save_txt}{T/F
File will be called *filename*.roleExperiment.}
}
\description{
writes to a .roleExperiment file that can be read into R using readRDs
}
