% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-divStats.R
\name{hillAbund}
\alias{hillAbund}
\alias{hillGenetic}
\alias{hillTrait}
\alias{hillPhylo}
\alias{richness}
\title{Hill number summary statistics for `roleData` objects}
\usage{
hillAbund(x, q = 1:4, ...)

hillGenetic(x, q = 1:4, ...)

hillTrait(x, q = 1:4, ...)

hillPhylo(x, q = 1:4, ...)

richness(x, ...)
}
\arguments{
\item{x}{the `roleData` object to calculate Hill numbers or richness from}

\item{q}{the Hill number exponents}

\item{...}{additional parameters, ignored}
}
\description{
These funcitons calculate numbers. The naming convention 
    indicates what type of hill number each function caculates; e.g. 
    `hillAbund` calcualtes abundance-based Hill numbers, while `hillTrait` 
    calculates trait-based Hill numbers
}
