% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-sumStats.R
\name{hillStats}
\alias{hillStats}
\alias{richness}
\title{Diversity summary statistics for `roleData` objects}
\usage{
hillStats(x, q, type, ...)

richness(x, ...)
}
\arguments{
\item{x}{the `roleData` object to calculate Hill numbers from}

\item{q}{the Hill number exponent}

\item{type}{Hill number type; one of 'abundance', 'trait', 'phylo'}

\item{...}{additional parameters, ignored}
}
\description{
Gets diversity summary statistics including Hill numbers
}
