% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalState.R
\name{getFinalState}
\alias{getFinalState}
\title{Get the final state of a model run}
\usage{
getFinalState(x, modID, ...)
}
\arguments{
\item{x}{the object to retrieve state from}

\item{modID}{optional argument to retrieve state for specific model(s) in a
`roleExperiment` object; can be a vector of unique IDs}
}
\description{
gets the final state of a model from either a `roleExperiment` 
    or a `roleModel`
}
\details{
if `modID` is not supplied and `x` is a `roleExperiment` then a 
    list of `roleData` objects for each model run is returned
}
