% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumStats.R
\name{getSumStats}
\alias{getSumStats}
\title{Get summary statistics for RoLE objects}
\usage{
getSumStats(x, funs, ...)
}
\arguments{
\item{x}{the object to calculate sum stats across}

\item{...}{additional named parameters passed to `fun`}

\item{fun}{the function(s) calculating the sum stats; can be a single 
function or a list of functions}
}
\description{
Applies different summary stats functions to `roleExperiment`,
    `roleModel`, or `roleData`
}
\details{
users can define their own functions, so long as they work on any
    object of class `roleData`
}
