% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{class}
\name{localComm-class}
\alias{localComm-class}
\alias{localComm}
\title{Local community}
\usage{
localComm(indSppTrt, indSeqs, sppGenDiv)
}
\description{
An S4 class to specify the state of a local community
}
\section{Slots}{

\describe{
\item{\code{indSppTrt}}{numeric matrix with rows for individuals; column 1 is species
ID (matches row number in `metaComm@sppData`); column 2 is trait value; 
column 3 is sequence (maybe? maybe shouldn't have it)}

\item{\code{indSeqs}}{character matrix with rows for individuals; column 1 is 
sequences for each individual}

\item{\code{sppGenDiv}}{a numeric matrix with rows for species; column 1 is species
level genetic diversity}
}}

