% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localComm.R
\name{localComm}
\alias{localComm}
\title{Specify a local community object}
\usage{
localComm(abundance, traits, pi)
}
\arguments{
\item{abundance}{a numeric vector of abundances for each species}

\item{traits}{matrix of traits; the first column specifies the species index
(i.e. the index of that species in the \code{abundance} vector) and the
second column specifies the trait value}

\item{pi}{a numeric vector of genetic diversities for each species}

\item{Smax}{a single integer specifying the total number of species ever
recorded in the local community (both locally extinct and extant)}
}
\value{
an object of class \code{localComm}
}
\description{
Specify a local community object
}
\details{
\code{Smax} is used for internal bookkeeping.  The dimensions of
\code{abundance}, \code{traits}, and \code{pi} can be greater than
\code{Smax}.  In such cases, \code{Smax} is used to index where to begin
adding new information (e.g. a new species can be added at the index
\code{Smax + 1}).
}
\seealso{
\code{\link{localComm-class}}
}
