% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{class}
\name{localComm-class}
\alias{localComm-class}
\alias{localComm}
\title{The local community of a `roleData`}
\usage{
localComm(indSpecies, indTrait, indSeqs)
}
\arguments{
\item{indSpecies}{indSpecies}

\item{indTrait}{indTrait}

\item{indSeqs}{indSeqs}

\item{spGenDiv}{spGenDiv}
}
\description{
An S4 class to specify the state of a local community
}
\section{Slots}{

\describe{
\item{\code{indSpecies}}{a numeric vector of the species IDs for each individual}

\item{\code{indTrait}}{a numeric vector of the trait values for each individual}

\item{\code{indSeqs}}{a character vector of the gene sequences for each individual}

\item{\code{spGenDiv}}{a numeric vector of the genetic diversities for each species}

\item{\code{spTrait}}{a numeric vector of the mean trait value of each species}

\item{\code{spAbund}}{a numeric vector of the abundance (number of individuals) of 
each species}

\item{\code{spAbundHarmMean}}{numeric vector of the harmonic mean of species 
abundances}

\item{\code{spLastOriginStep}}{numeric vector holding time of most recent origin of 
each species in the local community}

\item{\code{spExtinctionStep}}{numeric vector of most recent extirpation step of each 
species in the local community}

\item{\code{equilibProp}}{numeric proportion of equilibrium achived}
}}

