% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{class}
\name{metaComm-class}
\alias{metaComm-class}
\alias{metaComm}
\title{Meta community}
\usage{
metaComm(sppAbundTrt)
}
\description{
An S4 class to specify the state of a meta community
}
\section{Slots}{

\describe{
\item{\code{sppAbundTrt}}{numeric matrix with rows for species (row number is 
species ID); column 1 is relative abundance; column 2 is trait mean}
}}

