% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-sumStats.R
\name{rawAbundance}
\alias{rawAbundance}
\alias{rawSppID}
\alias{rawTraits}
\alias{rawGenDiv}
\alias{rawSeqs}
\alias{rawBranchLengths}
\alias{apePhylo}
\alias{rawRichness}
\alias{hillStats}
\title{Get raw data from `roleData` objects}
\usage{
rawAbundance(x, ...)

rawSppID(x, ...)

rawTraits(x, ...)

rawGenDiv(x, ...)

rawSeqs(x, ...)

rawBranchLengths(x, ...)

apePhylo(x, ...)

rawRichness(x, ...)

hillStats(x, q, type, ...)
}
\arguments{
\item{x}{the `roleData` object to calculate Hill numbers from}

\item{...}{additional parameters, ignored}

\item{q}{the Hill number exponent}

\item{type}{Hill number type; one of 'abundance', 'trait', 'phylo'}
}
\description{
Gets raw data

Gets Hill numbers
}
