% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData-rawStats.R
\name{rawAbundance}
\alias{rawAbundance}
\alias{rawSppID}
\alias{rawTraits}
\alias{rawGenDiv}
\alias{rawSeqs}
\alias{rawBranchLengths}
\alias{rawApePhylo}
\title{Get raw data from `roleData` objects}
\usage{
rawAbundance(x, ...)

rawSppID(x, ...)

rawTraits(x, ...)

rawGenDiv(x, ...)

rawSeqs(x, ...)

rawBranchLengths(x, ...)

rawApePhylo(x, ...)
}
\arguments{
\item{x}{the `roleData` object to get raw data from}

\item{...}{additional parameters, ignored}
}
\value{
a list with one element, inside that one element is a vector of the
    raw data
}
\description{
Gets raw data
}
\details{
These functions return a list of length one (with summary statistics 
    inside that list) because the summary stats themselves could be variable
    length. This is distinct from other summary stat functions which will
    always return a fixed length value.  When using raw summary stat 
    functions in `getSumStats` they will be returned in a list column
}
