% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleData.R
\docType{class}
\name{roleData-class}
\alias{roleData-class}
\alias{roleData}
\title{A collection of data representing one state of a roleModel.}
\usage{
roleData(localComm, metaComm, phylo)
}
\arguments{
\item{localComm}{A `localComm` object containing the local community.}

\item{metaComm}{A `metaComm` object containing the meta community.}

\item{phylo}{A `rolePhylo` object containing the model phylogeny.}
}
\description{
An S4 class to represent a roleModel state using `localComm`,
`metaComm`, & `rolephylo` bojects.
}
\section{Slots}{

\describe{
\item{\code{localComm}}{A `localComm` object containing the local community.}

\item{\code{metaComm}}{A `metaComm` object containing the meta community.}

\item{\code{phylo}}{A `rolePhylo` object containing the model phylogeny.}
}}

