% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleExperiment.R
\docType{class}
\name{roleExperiment-class}
\alias{roleExperiment-class}
\alias{roleExperiment}
\title{RoLE experiment}
\usage{
roleExperiment(allParams)
}
\description{
An S4 class to represent a complete self-enclosed experiment (a 
    model run or set of runs with associated meta data). Contains one or more 
    `roleModel` runs and the `roleParams` used to generate each of the runs
}
\section{Slots}{

\describe{
\item{\code{experimentMeta}}{data.frame of model meta data}

\item{\code{modelRuns}}{list of `roleData` objects containing outputs}

\item{\code{allParams}}{list of `roleParams` objects containing input params for 
each run}
}}

