% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleExperiment.R
\name{roleExperiment}
\alias{roleExperiment}
\title{Create a RoLE experiment}
\usage{
roleExperiment(allParams)
}
\arguments{
\item{allParams}{the list of model params to use for each model; can also be
`rolePriors` object}
}
\value{
a ready-to-run `roleExperiment`
}
\description{
Takes a list of parameter objects to initialize models for each
    parameter set.  The collection of models is seen as an experiment
}
