% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleModel.R
\docType{class}
\name{roleModel-class}
\alias{roleModel-class}
\title{An S4 class to specify the entire RoLE model}
\description{
An S4 class to specify the entire RoLE model
}
\section{Slots}{

\describe{
\item{\code{localComm}}{an object of class \code{localComm}}

\item{\code{metaComm}}{an object of class \code{metaComm}}

\item{\code{phylo}}{an object of class \code{rolePhylo}}

\item{\code{params}}{an object of calss \code{roleParams}}
}}

