% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleModel.R
\docType{class}
\name{roleModel-class}
\alias{roleModel-class}
\alias{roleModel}
\title{One run of the RoLE model}
\usage{
roleModel(params)
}
\description{
An S4 class to hold one model run of the RoLE model
}
\section{Slots}{

\describe{
\item{\code{modelSteps}}{a list of roleData objects, one for each recorded time step}

\item{\code{params}}{a roleParams object with the RoLE model params}
}}

