% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleModel.R
\name{roleModel}
\alias{roleModel}
\title{Create a RoLE model object}
\usage{
roleModel(localComm, metaComm, phylo, params)
}
\arguments{
\item{localComm}{an object of class \code{localComm}}

\item{metaComm}{an object of class \code{metaComm}}

\item{phylo}{an object of class \code{rolePhylo}}

\item{params}{an object of calss \code{roleParams}}
}
\value{
an object of class \code{roleModel}
}
\description{
Create a RoLE model object
}
