% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleSim.R
\name{roleSim}
\alias{roleSim}
\title{RoLE model simulation}
\usage{
roleSim(params = NULL, init = NULL, nstep = 100, nsim = 1)
}
\arguments{
\item{params}{list of parameters; if left unspecified a default set of "plausible"
parameter values is used}

\item{init}{an optional initial condition specified as a \code{roleModelCpp}
object}

\item{nstep}{number of simulation steps to run; if \code{prop_equilib} is
specified in the \code{params} list, \code{nstep} will over-ride that
parameter}

\item{nsim}{number of simulations to run}

\item{nout}{frequency of intermediate results output (not yet implemented)}
}
\description{
Simulate communities under the RoLE model. The key distinction
between the two functions is that \code{roleSim} is optimized to run many
simulations, while \code{roleSimPlay} is meant to run one simulation with
periodic output of results for visualization and exploration.
\code{roleSimPlay} is intended primarily for the accompanying Shiny App. R
}
\details{
Stub
}
