% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleSimPlots.R
\name{roleDistPlot}
\alias{roleDistPlot}
\alias{roleTSPlot}
\alias{roleDistAnim}
\alias{roleTSAnim}
\title{Plotting RoLE model simulations}
\usage{
roleDistPlot(x, type = all.types)

roleTSPlot(x, type = all.types)

roleDistAnim(roleSim, type = all.types)

roleTSAnim(roleSim, type = all.types)
}
\arguments{
\item{x}{a \code{roleSim} object}

\item{type}{a string specifying which type of data to plot}
}
\value{
A \code{plotly} object
}
\description{
Plot the dynamic distributions of RoLE model simulations and the
timeseries of RoLE model summary statitistics
}
\details{
Stub
}
