% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciation.R
\name{speciation}
\alias{speciation}
\title{Speciation}
\usage{
speciation(x, i, ...)
}
\arguments{
\item{x}{the object which determines method dispatch}

\item{i}{the index of the species undergoing speciation}
}
\description{
Generic and methods for computing the speciation process
}
